prd-pilot/backend/requirements.txt
fastapi==0.126.0
uvicorn[standard]==0.43.0
openai==1.55.0
httpx==0.28.1
fastmcp==2.13.1
mcp==1.25.0
pydantic==2.12.5
python-dotenv==1.2.2
pytest==8.4.2
pytest-asyncio==1.2.0
